Added command to validate a HelloCommand
Refactored error methods to return an owned string. This should make
it possible to return error strings that are created at runtime.

Corrected tests that checked against a wrong ID for the hello
command. The command is HELO to fit in the original 4 bytes for
a command ID. Now the field is 8 bytes but I did not adjust my
standard yet.