Started work on the PUT command
Added a new command and started working on tests. Not much right
now but I had to add a new crate to handle transforming binary vec
to a u32. This is because the sizes in my messages are in 32 bit
and rust only lets talk about sizes in term of the 'usize' type.

Changed the way I deal with the message ID, favoring static
values for the idents of the messages.