Added tests
Reworked the tests of the ObjProperties struct to use 4 byte sizes.
This is due to the new crate to serialize u32 to vectors and to
"safely" go between usize's and u32's.

Transformed the result_from_* functions to use a generic error
type.

Added tests for the remaining functions in the utils library.